useSWRを設定
