+++ src/component/template/Hero/index.tsx
@@ -1,5 +1,7 @@
+import axios from 'axios';
 import Link from 'next/link';
 import React from 'react';
+import useSWR from 'swr';
 import { storyList } from './data';
 import { Image, ImageProps } from '@/component/atom/Image';
 
@@ -13,14 +15,28 @@ export interface HeroProps {
   button: { text: string; link: string };
 }
 
-interface Props {
-  number: number;
+export interface GssProps {
+  タイムスタンプ?: string;
+  お名前?: string;
+  メールアドレス: string;
+  電話番号?: string;
+  プライバシーポリシーへの同意?: string;
 }
 
-export const Hero: React.FC<Props> = (props: Props) => {
-  const data = storyList.default;
-  data.number = props.number;
-  return <HeroPresenter {...data} />;
+export const Hero: React.FC = () => {
+  const value = storyList.default;
+  const fetcher = (url: string) => axios.get(url).then((res) => res.data);
+  const { data } = useSWR<GssProps[]>(
+    'https://api.steinhq.com/v1/storages/624e61c04906bb053738ccf4/sheet1',
+    fetcher,
+  );
+
+  //ユニーク数にするため、メールアドレスの重複を削除したuniqueListを作成
+  const list = data?.map((item) => item['メールアドレス']);
+  const set = new Set(list);
+  const uniqueList = Array.from(set);
+  value.number = uniqueList.length;
+  return <HeroPresenter {...value} />;
 };
 
 export const HeroPresenter: React.FC<HeroProps> = ({
