+++ src/pages/index.tsx
@@ -1,6 +1,3 @@
-import axios from 'axios';
-import { GetServerSideProps } from 'next';
-import { ReactNode } from 'react';
 import { Benefit } from '@/component/template/Benefit';
 import { Cta } from '@/component/template/Cta';
 import { Hero } from '@/component/template/Hero';
@@ -9,11 +6,11 @@ import { PageHeader } from '@/component/template/PageHeader';
 import { Price } from '@/component/template/Price';
 import { Sympathy } from '@/component/template/Sympathy';
 
-const Home: React.FC<Props> = (props: Props) => (
+const Home: React.FC = () => (
   <>
     <PageHeader />
     <main className='max-w-full'>
-      <Hero number={props.num} />
+      <Hero />
       <Sympathy />
       <Benefit />
       <Price />
@@ -24,32 +21,3 @@ const Home: React.FC<Props> = (props: Props) => (
 );
 
 export default Home;
-
-interface Props {
-  num: number;
-  children?: ReactNode;
-}
-
-interface GssProps {
-  タイムスタンプ?: string;
-  お名前?: string;
-  メールアドレス: string;
-  電話番号?: string;
-  プライバシーポリシーへの同意?: string;
-}
-
-const gssUrl = 'https://api.steinhq.com/v1/storages/624e61c04906bb053738ccf4/sheet1';
-
-export const getServerSideProps: GetServerSideProps = async () => {
-  const response = await axios.get(gssUrl);
-  const list = response.data.map((data: GssProps) => data['メールアドレス']);
-  //ユニーク数にするため、メールアドレスの重複を削除したuniqueListを作成
-  const set = new Set(list);
-  const uniqueList = Array.from(set);
-  const props: Props = {
-    num: uniqueList.length,
-  };
-  return {
-    props: props,
-  };
-};
